Extract parseTimeString util function
